requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
mysql-connector-python==8.1.0
requests==2.31.0
pytest==7.4.0
Flask-Bcrypt==1.0.0
Flask-RESTful==0.3.9